Move PerformanceContext.Scope to own file
